refactor: Move write to api/write.rs
